refactor: organizado o de resumo e shared
