+++ lib/src/perfil/components/new_tag_dialog.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:gastos/Model/Tag.dart';
-import 'package:gastos/ModelHelper/DatabaseHelper.dart';
+import 'package:gastos/src/shared/models/Tag.dart';
+import 'package:gastos/src/shared/repositories/DatabaseHelper.dart';
 
 void adicionarTag(String newTagName) {
   DatabaseHelper databaseHelper = DatabaseHelper();
